requirements.txt
grequests
beautifulsoup4
boto3
python-dotenv
dash
dash-bootstrap-components
pandas
plotly
schedule